refactor: 용도에 따른 디렉토리 구조 변경
