+++ src/App.js
@@ -1,12 +1,12 @@
-const { GAME_MESSAGES } = require("./constant");
-const GameInfo = require("./GameInfo");
-const InputView = require("./InputView");
-const OutputView = require("./OutputView");
+const { GAME_MESSAGES, INITIALIZE_VALUES } = require("./constants/constant");
+const GameInfo = require("./domain/GameInfo");
+const InputView = require("./ui/InputView");
+const OutputView = require("./ui/OutputView");
 
 class App {
   constructor() {
-    GameInfo.numberOfPlayGames = 0;
-    GameInfo.gameResult = "실패";
+    GameInfo.numberOfPlayGames = INITIALIZE_VALUES.zero;
+    GameInfo.gameResult = INITIALIZE_VALUES.failure;
   }
 
   play() {
