+++ src/domain/BridgeRandomNumberGenerator.js
