+++ src/domain/GameInfo.js
